+++ src/test/java/uk/gov/hmcts/reform/civil/handler/callback/camunda/notification/ChangeOfRepresentationNotificationHandlerTest.java
@@ -247,7 +247,7 @@ class ChangeOfRepresentationNotificationHandlerTest extends BaseCallbackHandlerT
     }
 
     @Nested
-    class SkipEmailNotification {
+    class SkipNotifyParties {
 
         @Test
         void shouldNotSendEmailWhenNotifyingFormerSolForLip() {
